mexi no banco de dados :(

+++ src/bean/LvbVendaVeiculos.java
@@ -1,8 +1,8 @@
 package bean;
-// Generated 25/11/2025 17:26:07 by Hibernate Tools 4.3.1
+// Generated 09/12/2025 15:44:41 by Hibernate Tools 4.3.1
 
 
-
+import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -23,16 +23,16 @@ import javax.persistence.Table;
 public class LvbVendaVeiculos  implements java.io.Serializable {
 
 
-     private int lvbIdVendaVeiculos;
+     private Integer lvbIdVendaVeiculos;
      private LvbVeiculos lvbVeiculos;
      private LvbVenda lvbVenda;
      private int lvbQuantidade;
-     private double lvbValotunitario;
+     private BigDecimal lvbValotunitario;
 
     public LvbVendaVeiculos() {
     }
 
-    public LvbVendaVeiculos(LvbVeiculos lvbVeiculos, LvbVenda lvbVenda, int lvbQuantidade, double lvbValotunitario) {
+    public LvbVendaVeiculos(LvbVeiculos lvbVeiculos, LvbVenda lvbVenda, int lvbQuantidade, BigDecimal lvbValotunitario) {
        this.lvbVeiculos = lvbVeiculos;
        this.lvbVenda = lvbVenda;
        this.lvbQuantidade = lvbQuantidade;
@@ -43,11 +43,11 @@ public class LvbVendaVeiculos  implements java.io.Serializable {
 
     
     @Column(name="lvb_id_venda_veiculos", unique=true, nullable=false)
-    public int getLvbIdVendaVeiculos() {
+    public Integer getLvbIdVendaVeiculos() {
         return this.lvbIdVendaVeiculos;
     }
     
-    public void setLvbIdVendaVeiculos(int lvbIdVendaVeiculos) {
+    public void setLvbIdVendaVeiculos(Integer lvbIdVendaVeiculos) {
         this.lvbIdVendaVeiculos = lvbIdVendaVeiculos;
     }
 
@@ -83,11 +83,11 @@ public class LvbVendaVeiculos  implements java.io.Serializable {
 
     
     @Column(name="lvb_valotunitario", nullable=false, precision=10)
-    public double getLvbValotunitario() {
+    public BigDecimal getLvbValotunitario() {
         return this.lvbValotunitario;
     }
     
-    public void setLvbValotunitario(double lvbValotunitario) {
+    public void setLvbValotunitario(BigDecimal lvbValotunitario) {
         this.lvbValotunitario = lvbValotunitario;
     }
